update dynamic pricing

+++ accommodation_service/src/main/java/huy/project/accommodation_service/infrastructure/repository/model/UnitPriceCalendarModel.java
@@ -54,7 +54,7 @@ public class UnitPriceCalendarModel extends AuditTable {
 
     // Occupancy-based pricing
     @Column(name = "base_occupancy")
-    BigDecimal baseOccupancy; // Standard occupancy this price is for
+    Integer baseOccupancy; // Standard occupancy this price is for
     @Column(name = "extra_person_fee")
     BigDecimal extraPersonFee; // Fee per additional guest
